Add onlyNew configuration parameters
This commit adds three new configuration parameters: onlyNewImages, onlyNewMetadata, and gracePeriodMins. The first two will prevent the script from downloading images and metadata older than the script's runtime, and the third is the amount of minutes to accept older images before the start of the script--in case it's down for some period of time.